feat: Updated migrations and added logic to update accounts and vouches table in historydb

+++ sequencer/db/historydb/migrations/001_init_historydb.sql
@@ -15,9 +15,9 @@ CREATE TABLE IF NOT EXISTS tx (
     batch_num BIGINT NOT NULL,
     position INT NOT NULL,
     type VARCHAR(40) NOT NULL,
-    from_idx BIGINT,
+    from_idx INTEGER,
     from_eth_addr BYTEA,
-    to_idx BIGINT NOT NULL,
+    to_idx INTEGER NOT NULL,
     to_eth_addr BYTEA,
     amount DECIMAL(78,0) NOT NULL,
     block_number BIGINT,
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS batch (
 -- Create account scheme
 CREATE TABLE IF NOT EXISTS account (
     item_id SERIAL,
-    idx BIGINT PRIMARY KEY,
+    idx INTEGER PRIMARY KEY,
     eth_addr BYTEA NOT NULL,
     balance DECIMAL(78,0) NOT NULL,
     score DECIMAL(78,0) NOT NULL,
@@ -47,9 +47,9 @@ CREATE TABLE IF NOT EXISTS account (
 -- Create vouch scheme
 CREATE TABLE IF NOT EXISTS vouch (
     idx BIGINT PRIMARY KEY,
-    from_idx BIGINT,
+    from_idx INTEGER,
     from_eth_addr BYTEA,
-    to_idx BIGINT,
+    to_idx INTEGER,
     to_eth_addr BYTEA
 );
